+++ CHANGELOG.md
@@ -9,8 +9,10 @@
 - WASM API
   - Renamed host function `trace` to `host_log`.
   - Added `host_log!` and `exec_command!` macros for working with host functions.
-  - Added `default_version` to `ToolMetadataOutput`.
+  - Added `default_version` field to `ToolMetadataOutput`.
+  - Added `home_dir` field to `LocateBinsInput`.
   - Updated `exec_command` with stream/inherit support.
+  - Updated `bin_path` in `LocateBinsOutput` and `ShimConfig` to a `PathBuf`.
 
 ## 0.13.1
