+++ crates/cli/tests/list_test.rs
@@ -1,13 +1,13 @@
 mod utils;
 
-use proto_core::Manifest;
+use proto_core::ToolManifest;
 use utils::*;
 
 #[test]
 fn lists_local_versions() {
     let temp = create_empty_sandbox();
 
-    let mut manifest = Manifest::load(temp.path().join("tools/node/manifest.json")).unwrap();
+    let mut manifest = ToolManifest::load(temp.path().join("tools/node/manifest.json")).unwrap();
     manifest.default_version = Some("19.0.0".into());
     manifest.installed_versions.insert("19.0.0".into());
     manifest.installed_versions.insert("18.0.0".into());
