+++ crates/cli/tests/plugins_test
