+++ crates/cli/tests/run_test.rs
@@ -1,6 +1,6 @@
 mod utils;
 
-use proto_core::Manifest;
+use proto_core::ToolManifest;
 use starbase_sandbox::predicates::prelude::*;
 use std::{env, fs};
 use utils::*;
@@ -181,7 +181,7 @@ fn updates_last_used_at() {
         .arg("--version")
         .assert();
 
-    let manifest = Manifest::load(&manifest_file).unwrap();
+    let manifest = ToolManifest::load(&manifest_file).unwrap();
     let last_used_at = manifest.versions.get("19.0.0").unwrap().last_used_at;
 
     assert!(last_used_at.is_some());
@@ -195,7 +195,7 @@ fn updates_last_used_at() {
         .arg("--version")
         .assert();
 
-    let manifest = Manifest::load(&manifest_file).unwrap();
+    let manifest = ToolManifest::load(&manifest_file).unwrap();
 
     assert_ne!(
         last_used_at,
