--- crates/core/README.md
@@ -1,5 +1,0 @@
-# proto_core
-
-![Crates.io](https://img.shields.io/crates/v/proto_core) ![Crates.io](https://img.shields.io/crates/d/proto_core)
-
-Core [proto](https://moonrepo.dev/proto) helpers and traits that all languages/tools must implement.
