+++ crates/core/src/tool.rs
