+++ crates/core/src/tool_manifest.rs
