+++ crates/core/src/tools_config.rs
