+++ crates/core/src/version.rs
