+++ crates/core/src/version_detector.rs
