+++ crates/core/src/version_resolver.rs
