+++ crates/core/templates/bash_global.tpl
