+++ crates/core/templates/bash_header.tpl
