+++ crates/core/templates/bash_local.tpl
