+++ crates/core/templates/cmd_global.tpl
