+++ crates/core/templates/pwsh_global.tpl
