+++ crates/core/templates/pwsh_header.tpl
