+++ crates/core/templates/pwsh_local.tpl
