+++ crates/core/tests/version_detector_test
