+++ crates/core/tests/version_test.rs
