--- crates/core2/Cargo.toml
@@ -1,36 +1,0 @@
-[package]
-name = "proto_core"
-version = "0.13.1"
-edition = "2021"
-license = "MIT"
-description = "Core proto APIs."
-homepage = "https://moonrepo.dev/proto"
-repository = "https://github.com/moonrepo/proto"
-
-[dependencies]
-proto_pdk_api = { version = "0.3.1", path = "../pdk-api" }
-proto_wasm_plugin = { version = "0.3.1", path = "../wasm-plugin" }
-warpgate = { version = "0.2.0", path = "../warpgate" }
-cached = { workspace = true }
-extism = { workspace = true }
-fs4 = { version = "0.6.6", features = ["sync"] }
-human-sort = { workspace = true }
-miette = { workspace = true }
-once_cell = { workspace = true }
-regex = { workspace = true }
-reqwest = { workspace = true }
-rustc-hash = { workspace = true }
-semver = { workspace = true }
-serde = { workspace = true }
-serde_json = { workspace = true }
-sha2 = { workspace = true }
-starbase_archive = { workspace = true }
-starbase_styles = { workspace = true }
-starbase_utils = { workspace = true }
-thiserror = { workspace = true }
-tinytemplate = "1.2.1"
-tracing = { workspace = true }
-url = "2.4.0"
-
-[dev-dependencies]
-starbase_sandbox = { workspace = true }
