+++ crates/wasm-plugin/tests/wasm_plugin_test
